patch: finished first version of the project [Task]: test_task-7
[Description list]
build: finished first version of the project

[Notes]
<notes>
