--- images_logger
@@ -1,22 +1,0 @@
-Converted image - images/gray_test_image_98.jpg; time - 2024-05-07 18:15:51.572577
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:15:51.646892
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:19:18.410512
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:19:18.480436
-Converted image - <coroutine object process_uploaded_file_async at 0x7e8f2548d540>; time - 2024-05-07 18:20:15.488747
-Converted image - <coroutine object process_uploaded_file_async at 0x7e8f2548e040>; time - 2024-05-07 18:20:15.863608
-Converted image - images/gray_test_image_98.jpg; time - 2024-05-07 18:23:40.280349
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:23:40.349298
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:24:45.846815
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:24:45.919015
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:28:54.260600
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:28:54.335661
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:34:05.742156
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:34:05.814446
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:34:48.752784
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:34:48.824177
-Converted image - images/gray_test_image_98.jpg; time - 2024-05-07 18:35:15.999276
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:35:16.071490
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:36:24.680049
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:36:24.747574
-Converted image - images/gray_test_image_98.jpg; time - 2024-05-07 18:36:41.255241
-Returning cached image for test_image_98.jpg; time - 2024-05-07 18:36:41.324627
